Improve post admin (update)
The javascript will no longer generate html itself (except for forms).
The hole post HTML is generated in a single SLIM template file (_show.html.slim)
It also centralize the generation of the post's HTML for index and show (using
locals in order to choose the long version for the action show and the short one
for the index action.
The ajax does not parse the JSON anymore but only replace the form with the HTML
fetch from the update.

- add a partial to display the post (in both short and long format)
  - _show.html.slim
  - add a local "short" to be able from the browser to know if its a long or a
    short post
- update a post will fetch the html (without layout)
  - template on server side generate the HTML (centralize the generation of the
    html into a single SLIM template file)
  - short information is sent in the ajax request in order to require the
    template to render short_html or html (full).
  - show can render no layout (if request.xhr)
  - update can render show without layout (for html replacement) if request.xhr